% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaesquejars-package.R
\docType{package}
\name{kafkaesquejars}
\alias{kafkaesquejars}
\title{Java Archive Wrapper Supporting the 'kafkaesque' Package}
\description{
Kafka is a big data technology providing a scalable, persistent, distributed,
high throughput event log to which clients can subscribe topics or publish
messages.
}
\author{
Peter Meissner \href{mailto:retep.meissner@gmail.com}{retep.meissner@gmail.com}
}
